refactor: pr 수정
리뷰 점수 double->int
입력 예외 NOT_FOUND->BAD_REQUEST
제목 길이 10에서 30으로 수정
리뷰 삭제할 때 이미지 삭제도 추가

+++ src/main/kotlin/sosteam/deamhome/domain/review/handler/response/ReviewResponse.kt
@@ -6,7 +6,7 @@ data class ReviewResponse(
 	val reviewId: String,
 	val title: String,
 	val content: String,
-	val score: Double,
+	val score: Int,
 	val status: Boolean,
 	val userId: String,
 	val itemId: String,
